back/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-types==0.7.0",
    "antlr4-python3-runtime==4.9.3",
    "anyio==4.8.0",
    "black==25.1.0",
    "click==8.1.8",
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "eva-decord==0.6.1",
    "fastapi==0.115.10",
    "filelock==3.17.0",
    "fonttools==4.56.0",
    "fsspec==2025.2.0",
    "h11==0.14.0",
    "hydra-core==1.3.2",
    "idna==3.10",
    "iniconfig==2.0.0",
    "iopath==0.1.10",
    "isort==6.0.1",
    "jinja2==3.1.5",
    "kiwisolver==1.4.8",
    "markupsafe==3.0.2",
    "matplotlib==3.10.1",
    "mpmath==1.3.0",
    "mypy==1.15.0",
    "mypy-extensions==1.0.0",
    "networkx==3.4.2",
    "numpy==2.2.3",
    "nvidia-cublas-cu12==12.4.5.8",
    "nvidia-cuda-cupti-cu12==12.4.127",
    "nvidia-cuda-nvrtc-cu12==12.4.127",
    "nvidia-cuda-runtime-cu12==12.4.127",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12==11.2.1.3",
    "nvidia-curand-cu12==10.3.5.147",
    "nvidia-cusolver-cu12==11.6.1.9",
    "nvidia-cusparse-cu12==12.3.1.170",
    "nvidia-cusparselt-cu12==0.6.2",
    "nvidia-nccl-cu12==2.21.5",
    "nvidia-nvjitlink-cu12==12.4.127",
    "nvidia-nvtx-cu12==12.4.127",
    "omegaconf==2.3.0",
    "opencv-python==4.11.0.86",
    "packaging==24.2",
    "pathspec==0.12.1",
    "pillow==11.1.0",
    "platformdirs==4.3.6",
    "pluggy==1.5.0",
    "portalocker==3.1.1",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pyparsing==3.2.1",
    "pytest==8.3.4",
    "python-dateutil==2.9.0.post0",
    "pyyaml==6.0.2",
    "sam-2",
    "setuptools==75.8.2",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.46.0",
    "sympy==1.13.1",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "tqdm==4.67.1",
    "triton==3.2.0",
    "typing-extensions==4.12.2",
    "uvicorn==0.34.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.0",
    "mypy>=1.15.0",
    "pytest>=8.3.4",
]

[tool.uv.sources]
sam-2 = { path = "submodules/sam2" }
